Добавил класс Game
